backend: edit structure
