+++ src/routes/auth/jwt.stategy.ts
@@ -1,7 +1,7 @@
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
-import { JwtUser } from "../common/types/jwtuser.type";
+import { JwtUser } from "../../types/jwtuser.type";
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
