+++ src/routes/vote/vote.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { VoteController } from "./vote.controller";
 import { VoteService } from "./vote.service";
-import { VoteDBModule } from "../common/db_services/votes/voteDB.module";
-import { UserDBModule } from "../common/db_services/users/userDB.module";
-import { MovieDBModule } from "../common/db_services/movies/movieDB.module";
-import { PrismaService } from "../common/db_services/prisma.service";
+import { VoteDBModule } from "../../common/db_services/votes/voteDB.module";
+import { UserDBModule } from "../../common/db_services/users/userDB.module";
+import { MovieDBModule } from "../../common/db_services/movies/movieDB.module";
+import { PrismaService } from "../../common/db_services/prisma.service";
 
 @Module({
   imports: [VoteDBModule, UserDBModule, MovieDBModule],
