+++ src/routes/vote/vote.service.ts
@@ -1,7 +1,7 @@
 import { ConflictException, Injectable, InternalServerErrorException } from "@nestjs/common";
-import { MovieDBService } from "../common/db_services/movies/movieDB.service";
-import { UserDBService } from "../common/db_services/users/userDB.service";
-import { VoteDBService } from "../common/db_services/votes/voteDB.service";
+import { MovieDBService } from "../../common/db_services/movies/movieDB.service";
+import { UserDBService } from "../../common/db_services/users/userDB.service";
+import { VoteDBService } from "../../common/db_services/votes/voteDB.service";
 import { Movie, Prisma } from "@prisma/client";
 
 @Injectable()
