+++ src/types/user.dto/password_new.dto.ts
